functions/requirements.txt
praw==7.1.0
monkeylearn==3.5.2
Unidecode==1.2.0
Markdown==3.3.4
fuzzywuzzy==0.18.0
pyfunctional==1.4.3
google_api_python_client==2.6.0
IMDbPY==2021.4.18
python-dotenv==0.17.1
levenshtein=0.12.0